Add capability to be used by anyone
Anyone having the right (DAC+MAC) to modify a file
can use vi to do that. To keep the Smack labels of
the edited file, the vim process must have the
CAP_MAC_ADMIN capability. But not much user session
have it. That is why the CAP_MAC_ADMIN capability
is set to the executable files using setcap at install.

Change-Id: I6df60aaa4ef92d539320a8a51dce9dffce310a50
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>